+++ sites/postcss-preset-env/postcss.config.cjs
@@ -1,10 +1,15 @@
+const crypto = require('crypto');
+const cwd = process.cwd();
+
 module.exports = ctx => {
 	const isProd = ctx.env === 'production';
 
 	return {
 		map: !isProd,
 		plugins: {
-			'postcss-import': {},
+			'postcss-import': {
+				nameLayer: hashLayerName,
+			},
 			'postcss-preset-env': {
 				stage: 0,
 				preserve: true,
@@ -21,3 +26,18 @@ module.exports = ctx => {
 		},
 	};
 };
+
+function hashLayerName(index, rootFilename) {
+	if (!rootFilename) {
+		return `import-anon-layer-${index}`;
+	}
+
+	// A stable, deterministic and unique layer name:
+	// - layer index
+	// - relative rootFilename to current working directory
+	return `import-anon-layer-${crypto
+		.createHash('sha256')
+		.update(`${index}-${rootFilename.split(cwd)[1]}`)
+		.digest('hex')
+		.slice(0, 12)}`;
+}