+++ plugins/postcss-cascade-layers/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to PostCSS Cascade Layers
 
+### Unreleased
+
+- Fix broken `@keyframes` in `@layer`.
+- Add support for `@container` as a conditional rule.
+
 ### 1.0.5 (July 8, 2022)
 
 - Fix case insensitive `@layer` matching (`@LaYeR`).